Use version 2 of the API on api ls commands
Groups property is being used on:
- portfolios
- models
- data_files
- analyses

But v1 of the API does not return property groups. Only v2 does.

After checking the swagger, updating to v2 seems better than removing groups from the sanitization of the returned body.

Using API v2 for:
- /models/${PIWIND_VERSION}/chunking_configuration
- /models/${PIWIND_VERSION}/scaling_configuration

version 1 of the API lacks of these methods